requirements.txt
absl-py
astor==0.8.1
certifi==2020.12.5
coverage
Cython
gast
grpcio
h5py 
importlib-metadata
Keras-Applications
Keras-Preprocessing
Markdown
mkl-fft==1.2.1
mkl-random==1.1.1
mkl-service==2.3.0
numpy
olefile==0.46
Pillow
protobuf==3.14.0
scipy
six
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
tqdm
Werkzeug
zipp
